move navbar components into separate folder
moved from /pages/nav/*  -->  /navbar/*
