+++ src/components/navbar/GroomerNav.js
@@ -3,8 +3,6 @@ import { NavLink } from 'react-router-dom';
 import { Layout, Menu, Button, Row, Typography } from 'antd';
 import { useOktaAuth } from '@okta/okta-react';
 
-import './Navbar.css';
-
 import {
   UserOutlined,
   HomeOutlined,
@@ -33,11 +31,6 @@ function GroomerNav(props) {
         height: '100vh',
       }}
     >
-      <div className="logo">
-        <Title className="welcome" level={4}>
-          Express Groomers
-        </Title>
-      </div>
       <Menu
         theme="dark"
         onClick={props.handleClick}
