--- src/components/pages/Nav/Header.scss
@@ -1,6 +1,0 @@
-// @import "../../index.scss";
-.header, .avatar-item {
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-}
