--- src/components/pages/Nav/Navbar.css
@@ -1,5 +1,0 @@
-.logo {
-    height: 32px;
-    background: rgba(255, 255, 255, 0.2);
-    margin: 16px;
-  }
