+++ src/index.js
@@ -23,9 +23,9 @@ import { LoginPage } from './components/pages/Login';
 import { config } from './utils/oktaConfig';
 import { LoadingComponent } from './components/common';
 import { ProfilePages } from './components/pages/ProfilePage/';
-import Navbar from './components/pages/Nav/Navbar';
-import Header from './components/pages/Nav/Header';
-import Footer from './components/pages/Nav/Footer';
+import Navbar from './components/navbar/Navbar';
+import Header from './components/navbar/Header';
+import Footer from './components/navbar/Footer';
 
 ReactDOM.render(
   <Provider store={store}>
